refactor: adds more structure to the project with /education 'fagveven' subdirectories
