+++ src/actions/education/courses/create.ts
@@ -1,6 +1,6 @@
 'use server'
 import { safeServerCall } from '@/actions/safeServerCall'
-import type { CreateCourseTypes } from '@/services/courses/validation'
+import type { CreateCourseTypes } from '@/education/courses/validation'
 
 export async function createCourseAction(rawdata: FormData | CreateCourseTypes['Type']) {
     console.log('createCourseAction', rawdata)
