+++ src/actions/education/schools/destroy.ts
@@ -2,7 +2,7 @@
 import { createActionError } from '@/actions/error'
 import { safeServerCall } from '@/actions/safeServerCall'
 import { getUser } from '@/auth/getUser'
-import { destroySchool } from '@/services/schools/destroy'
+import { destroySchool } from '@/services/education/schools/destroy'
 import type { ActionReturn } from '@/actions/Types'
 
 export async function destroySchoolAction(id: number): Promise<ActionReturn<void>> {
