+++ src/app/admin/courses/page.tsx
@@ -1,0 +1,9 @@
+import PageWrapper from '@/app/_components/PageWrapper/PageWrapper'
+
+export default function CoursesAdmin() {
+    return (
+        <PageWrapper title="Emnekatalog">
+            <h1>Emnene</h1>
+        </PageWrapper>
+    )
+}
