+++ src/app/admin/schools/[shortname]/page.tsx
@@ -1,7 +1,7 @@
 import styles from './page.module.scss'
 import UpdateSchool from './UpdateSchool'
-import { destroySchoolAction } from '@/actions/schools/destroy'
-import { readSchoolAction } from '@/actions/schools/read'
+import { destroySchoolAction } from '@/education/schools/destroy'
+import { readSchoolAction } from '@/education/schools/read'
 import Form from '@/app/_components/Form/Form'
 import PageWrapper from '@/components/PageWrapper/PageWrapper'
 import School from '@/components/School/School'
