+++ src/app/education/page.tsx
@@ -1,0 +1,13 @@
+import PageWrapper from '@/components/PageWrapper/PageWrapper'
+
+export default function education() {
+    return (
+        <PageWrapper title="Fagveven">
+            <p>
+                Velkommen til omegas fagvev! Her finner du nyttig info om emner og utveksling skrevet
+                av studenter for studenter. Du kan også bidra selv ved å skrive om dine erfaringer
+                fra utveksling eller emner du har tatt.
+            </p>
+        </PageWrapper>
+    )
+}
