+++ src/app/education/schools/page.tsx
@@ -1,0 +1,10 @@
+import PageWrapper from '@/components/PageWrapper/PageWrapper'
+
+
+export default function Schools() {
+    return (
+        <PageWrapper title="Skoler">
+            <p>Dette er en liste skole. Du er velkommen til å legge til flere</p>
+        </PageWrapper>
+    )
+}
