+++ src/services/education/schools/ConfigVars.ts
