+++ src/services/education/schools/Types.ts
