+++ src/services/education/schools/destroy.ts
