+++ src/services/education/schools/read.ts
