+++ tsconfig.json
@@ -34,6 +34,7 @@
       "@/utils/*": ["src/utils/*"],
       "@/jwt/*": ["src/jwt/*"],
       "@/api/*": ["src/app/api/*"],
+      "@/education/*": ["src/actions/education/*", "src/services/education/*", "src/app/education/*"],
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
